requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==45.0.5
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20241021
distro==1.9.0
Django==5.1.4
django-cleanup==9.0.0
django-crispy-forms==2.3
django-import-export==4.3.3
django-meta==2.5.0
fonttools==4.55.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
oauthlib==3.3.1
openai==2.2.0
packaging==24.2
phonenumbers==9.0.19
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.12.0
pydantic_core==2.41.1
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.15.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.2
tablib==3.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3.post1
